fix validation in transfer dialog
