+++ src/modules/edc-demo/components/asset-editor-dialog/model/http-datasource-auth-header-type.ts
@@ -1,0 +1,1 @@
+export type HttpDatasourceAuthHeaderType = 'None' | 'Value' | 'Vault-Secret';
